update to ai controller and movement relationship
ai movement was running its own update. now interface it was already implementing has a OnTick() function that is implemented and called from the interface object instantiated inside of the ai controller.

also changed the name of the Target object in the patrol and chase components from moveref to movementController because it didn't really make any sense

            aiController.chaseComponentObject.OnTargetReachedListener;